+++ test/mocks/MockLayerWiseTransformer.h
@@ -5,6 +5,8 @@
 #ifndef RAMPACK_MOCKLAYERWISETRANSFORMER_H
 #define RAMPACK_MOCKLAYERWISETRANSFORMER_H
 
+#include <optional>
+
 #include <catch2/trompeloeil.hpp>
 
 #include "core/lattice/LayerWiseTransformer.h"
@@ -13,7 +15,7 @@
 class MockLayerWiseTransformer : public LayerWiseTransformer {
 protected:
     MAKE_CONST_MOCK2(transformShape, Shape(const Shape &shape, std::size_t layerIndex), override);
-    MAKE_CONST_MOCK0(getRequestedNumOfLayers, std::size_t(), override);
+    MAKE_CONST_MOCK0(getRequestedNumOfLayers, std::optional<std::size_t>(), override);
 
 public:
     explicit MockLayerWiseTransformer(LatticeTraits::Axis axis) : LayerWiseTransformer(axis) { }