+++ CHANGELOG.md
@@ -23,6 +23,8 @@ library (later, with Python bindings as well).
   [interaction center types](docs/shapes.md#interaction-center-types) of sphere-based shapes.
 * Added delayed and strided filling to [class `serial`](docs/initial-arrangement.md#class-serial) lattice populator
   via `start_from` and `every`. `n_shapes` is now optional and defaults to traversing the lattice until the end.
+* Added `cumulative` rotation mode and `turn_fraction(...)` angle specification to
+  [class `layer_rotate`](docs/initial-arrangement.md#class-layer_rotate).
 
 ### Changed
 
@@ -33,6 +35,8 @@ library (later, with Python bindings as well).
   This includes [class `polysphere`](docs/shapes.md#class-polysphere) itself, as well as the previously hard-only
   [class `polysphere_lollipop`](docs/shapes.md#class-polysphere_lollipop) and
   [class `polysphere_wedge`](docs/shapes.md#class-polysphere_wedge).
+
+
 ## [1.2.0] - 2023-12-03
 
 ### Fixed